refactor: move AbstractFlashcardViewer related methods to its file
removes a circular dependency

removes some unnecessary imports

makes easier to remove AbstractFlashcardViewer when the old reviewer is gone
